FPO-154: Fix app not actually starting for various reasons
The app was reliant on .env files in production, despite environment
variables being injected through ECS. I have disabled the check for these
files in production.

The app also didn't tell us if the server was running, so I have added
logging via winston and a callback on app.listen to tell us this.

I've also added some more constants, such as a constant for the server
port.

The healthcheck controllers were reliant on a REVISION file, which may
not exist in a development environment so basic ping/testing of whether
the server is up wasn't easy. I've added a basic promise to reject with
the value 'development' if it can't find these files.